+++ frontend/src/views/Admin.vue
@@ -6,6 +6,7 @@
       <header>
         <h3>Meny</h3>
       </header>
+      <Search v-model="searchTerm" @input="search"></Search>
       <EditableDishList></EditableDishList>
     </section>
 
@@ -15,28 +16,31 @@
 </template>
 
 <script>
+import axios from "axios";
 import TheHeader from "../components/TheHeader.vue";
 import Greeting from "../components/admin/Greeting.vue";
+import Search from "../components/Search";
 import EditableDishList from "../components/admin/EditableDishList.vue";
 import AddNewDish from "../components/admin/AddNewDish.vue";
 import AddAdmin from "../components/admin/AddAdmin.vue";
-import axios from "axios";
 
 export default {
   name: "AdminHome",
   components: {
     TheHeader,
     Greeting,
+    Search,
     EditableDishList,
     AddNewDish,
     AddAdmin
   },
   data() {
     return {
+      searchTerm: undefined,
       user: {}
     };
   },
-  created() {
+  beforeMount() {
     if (document.cookie) {
       let cookies = document.cookie.split(";");
       if (cookies.length > 1) {
@@ -55,6 +59,28 @@ export default {
       }
     } else {
       this.$router.push("/login");
+    }
+  },
+  methods: {
+    search() {
+      const webAPIUrl = `https://localhost:5001/api/dishes/`;
+      axios.get(webAPIUrl).then(response => {
+        let dishes = [];
+        response.data.forEach(dish => {
+          if (
+            dish.name.toLowerCase().includes(this.searchTerm.toLowerCase()) ||
+            dish.type.toLowerCase().includes(this.searchTerm.toLowerCase()) ||
+            dish.description
+              .toLowerCase()
+              .includes(this.searchTerm.toLowerCase())
+          ) {
+            dishes.push(dish);
+            this.dishes = dishes;
+          } else {
+            this.dishes = dishes;
+          }
+        });
+      });
     }
   }
 };
