+++ frontend/src/views/Homepage.vue
@@ -5,14 +5,20 @@
     <header class="dish-list-header">
       <h2 class="dish-list-heading">Meny</h2>
     </header>
-    <DishList class="menu"></DishList>
+    <section class="menu">
+      <Search v-model="searchTerm" @input="search"></Search>
+      <DishList :dishes="dishes"></DishList>
+    </section>
+
     <TheFooter></TheFooter>
   </main>
 </template>
 
 <script>
+import axios from "axios";
 import TheHeader from "../components/TheHeader.vue";
 import DishHighlight from "../components/user/DishHighlight.vue";
+import Search from "../components/Search";
 import DishList from "../components/user/DishList.vue";
 import TheFooter from "../components/TheFooter.vue";
 
@@ -21,8 +27,43 @@ export default {
   components: {
     TheHeader,
     DishHighlight,
+    Search,
     DishList,
     TheFooter
+  },
+  data() {
+    return {
+      searchTerm: undefined,
+      dishes: []
+    };
+  },
+  created() {
+    const webAPIUrl = `https://localhost:5001/api/dishes`;
+    axios.get(webAPIUrl).then(response => {
+      this.dishes = response.data;
+    });
+  },
+  methods: {
+    search() {
+      const webAPIUrl = `https://localhost:5001/api/dishes/`;
+      axios.get(webAPIUrl).then(response => {
+        let dishes = [];
+        response.data.forEach(dish => {
+          if (
+            dish.name.toLowerCase().includes(this.searchTerm.toLowerCase()) ||
+            dish.type.toLowerCase().includes(this.searchTerm.toLowerCase()) ||
+            dish.description
+              .toLowerCase()
+              .includes(this.searchTerm.toLowerCase())
+          ) {
+            dishes.push(dish);
+            this.dishes = dishes;
+          } else {
+            this.dishes = dishes;
+          }
+        });
+      });
+    }
   }
 };
 </script>
@@ -57,7 +98,13 @@ export default {
 }
 
 .menu {
-  padding-bottom: 100px;
+  width: 100%;
+  max-width: 1000px;
+  margin: 0 auto;
+  padding: 0 20px 60px 20px;
+  display: grid;
+  grid-template-columns: 1fr;
+  gap: 40px;
 }
 
 @media only screen and (min-width: 700px) {
